+++ src/processors/scaler_audio_processor.py
@@ -7,7 +7,7 @@ import numpy.typing as npt
 
 from .abstract_data_processor import AbstractFitDataProcessor
 
-def reshape_to_transform(data : np.ndarray) -> npt.NDArray[np.float32]:
+def reshape_to_transform(data : np.ndarray) -> npt.NDArray[np.float64]:
     return np.expand_dims(data.flatten(), axis = 1)
 
 class MeanStandardScaler(StandardScaler):
@@ -30,13 +30,13 @@ class ScalerAudioProcessor(AbstractFitDataProcessor):
         self.mms = MinMaxScaler(feature_range=feature_range)
         super().__init__()
 
-    def forward(self, data: np.ndarray) -> npt.NDArray[np.float32]:
+    def forward(self, data: np.ndarray) -> npt.NDArray[np.float64]:
         shape = data.shape
         data = self.ssc.transform(reshape_to_transform(data)).reshape(shape)
         data = self.mms.transform(reshape_to_transform(data)).reshape(shape)
         return data
     
-    def backward(self, data: np.ndarray) -> npt.NDArray[np.float32]:
+    def backward(self, data: np.ndarray) -> npt.NDArray[np.float64]:
         shape = data.shape
         data = self.mms.inverse_transform(reshape_to_transform(data)).reshape(shape)
         data = self.ssc.inverse_transform(reshape_to_transform(data)).reshape(shape)
